Rerouted calendar to Timetable and removed the irrelevant imports
