+++ app/Response.hs
@@ -1,7 +1,6 @@
 module Response (module X) where
 
 import Response.About as X
-import Response.Calendar as X
 import Response.Draw as X
 import Response.Generate as X
 import Response.Image as X
