+++ app/Routes.hs
@@ -5,7 +5,7 @@ import Control.Monad (MonadPlus (mplus), msum)
 import Control.Monad.IO.Class (liftIO)
 import Controllers.Course as CoursesController (retrieveCourse, index, courseInfo, depts)
 import Controllers.Graph as GraphsController
-import Controllers.Timetable as GridController (gridResponse, exportTimetableImageResponse, exportTimetablePDFResponse)
+import Controllers.Timetable as TimetableController
 import Data.Text.Lazy (Text)
 import Database.CourseInsertion (saveGraphJSON)
 import Database.CourseQueries (getGraphJSON, retrievePost)
@@ -34,7 +34,6 @@ import Response
       generateResponse,
       postResponse,
       loadingResponse,
-      calendarResponse,
       graphImageResponse)
 
 routeResponses :: String -> Text -> Text -> ServerPartT IO Response
@@ -46,13 +45,13 @@ routeResponses staticDir aboutContents privacyContents =
 
 strictRoutes :: Text -> Text -> [ (String, ServerPart Response)] 
 strictRoutes aboutContents privacyContents = [
-    ("grid", GridController.gridResponse),
+    ("grid", TimetableController.gridResponse),
     ("graph", GraphsController.graphResponse),
     ("graph-generate", do method PUT
                           GraphsController.findAndSavePrereqsResponse),
     ("image", look "JsonLocalStorageObj" >>= graphImageResponse),
-    ("timetable-image", lookText' "session" >>= \session -> look "courses" >>= GridController.exportTimetableImageResponse session),
-    ("timetable-pdf", look "courses" >>= \courses -> look "JsonLocalStorageObj" >>= GridController.exportTimetablePDFResponse courses),
+    ("timetable-image", lookText' "session" >>= \session -> look "courses" >>= TimetableController.exportTimetableImageResponse session),
+    ("timetable-pdf", look "courses" >>= \courses -> look "JsonLocalStorageObj" >>= TimetableController.exportTimetablePDFResponse courses),
     ("post", retrievePost),
     ("post-progress", postResponse),
     ("draw", drawResponse),
@@ -67,7 +66,7 @@ strictRoutes aboutContents privacyContents = [
     ("courses", CoursesController.index),
     ("course-info", lookText' "dept" >>= CoursesController.courseInfo),
     ("depts", CoursesController.depts),
-    ("calendar", look "courses" >>= calendarResponse),
+    ("calendar", look "courses" >>= TimetableController.calendarResponse),
     ("loading", lookText' "size" >>= loadingResponse),
     ("save-json", lookBS "jsonData" >>= \jsonStr -> lookText' "nameData" >>= \nameStr -> liftIO $ saveGraphJSON jsonStr nameStr)
     ]
