+++ courseography.cabal
@@ -88,7 +88,6 @@ executable courseography
     MasterTemplate,
     Response,
     Response.About,
-    Response.Calendar,
     Response.Draw,
     Response.Image,
     Response.Loading,
